chore(web-api): separate the telemetry-configuration into aspnet package
