+++ src/Digdir.Library.Utils.AspNet/Digdir.Library.Utils.AspNet.csproj
@@ -7,9 +7,14 @@
 
     <ItemGroup>
         <PackageReference Include="Azure.Identity" Version="1.13.0" />
+        <PackageReference Include="Azure.Monitor.OpenTelemetry.AspNetCore" Version="1.2.0" />
+        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.22.0" />
         <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.1" />
         <PackageReference Include="AspNetCore.HealthChecks.UI.Client" Version="8.0.1" />
         <PackageReference Include="Microsoft.Azure.AppConfiguration.AspNetCore" Version="8.0.0" />
+        <PackageReference Include="Npgsql.OpenTelemetry" Version="8.0.5" />
+        <PackageReference Include="OpenTelemetry" Version="1.9.0" />
+        <PackageReference Include="OpenTelemetry.Instrumentation.Runtime" Version="1.9.0" />
     </ItemGroup>
 
 </Project>
